src/templates/packages/back/common/src/@api/common/country/resolvers/common-create-countries.resolver.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { ICommandBus, IQueryBus } from 'aurora-ts-core';

// custom items
import { CommonCreateCountriesResolver } from './common-create-countries.resolver';
import { countries } from '../../../../@apps/common/country/infrastructure/seeds/country.seed';
import { CommonCreateCountryInput } from './../../../../graphql';

describe('CommonCreateCountriesResolver', () =>
{
    let resolver: CommonCreateCountriesResolver;
    let queryBus: IQueryBus;
    let commandBus: ICommandBus;

    beforeAll(async () =>
    {
        const module: TestingModule = await Test.createTestingModule({
            providers: [
                CommonCreateCountriesResolver,
                {
                    provide : IQueryBus,
                    useValue: {
                        ask: () => { /**/ },
                    }
                },
                {
                    provide : ICommandBus,
                    useValue: {
                        dispatch: () => { /**/ },
                    }
                },
            ]
        }).compile();

        resolver    = module.get<CommonCreateCountriesResolver>(CommonCreateCountriesResolver);
        queryBus    = module.get<IQueryBus>(IQueryBus);
        commandBus  = module.get<ICommandBus>(ICommandBus);
    });

    test('CommonCreateCountriesResolver should be defined', () =>
    {
        expect(resolver).toBeDefined();
    });

    describe('main', () =>
    {
        test('CommonCreateCountriesResolver should be defined', () =>
        {
            expect(resolver).toBeDefined();
        });

        test('should return an countries created', async () =>
        {
            expect(await resolver.main(<CommonCreateCountryInput[]>countries)).toBe(true);
        });
    });
});